Refactor selection to be inside numba
